Finalden Oncesi 4

+++ src/main/java/com/example/ankaref/DataAccess/EventRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EventRepository extends JpaRepository<Event,Integer> {//Evet olduğu Data yapılacak işlemlerin olduğu kısım
+public interface EventRepository extends JpaRepository<Event,Long> {//Evet olduğu Data yapılacak işlemlerin olduğu kısım
 
 }
